chapter2-extracting-data-from-pdfs/extracting-text-from-images-in-pdf/requirements.txt
PyMuPDF
Pillow
pytesseract